Store the full 64-bit of the file size, not just the lower 32-bit. (Fixes issue #265) : File size displayed as 142MB for files 8.5GB in size

+++ trunk/src/SearchInfo.h
@@ -31,7 +31,7 @@ public:
     ~CSearchInfo(void);
 
     std::wstring                filepath;
-    DWORD                       filesize;
+    __int64                     filesize;
     std::vector<DWORD>          matchlinesnumbers;
     std::vector<std::wstring>   matchlines;
     CTextFile::UnicodeType      encoding;
